pyproject.toml
[tool.poetry]
name = "invert-image"
version = "0.1.0"
description = "A CWL workflow to invert an image"
authors = ["Steven Gillies <steven.gillies@telespazio.com>"]
license = "Apache 2.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^10.1.0"


[tool.poetry.group.dev.dependencies]
black = "^23.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
